CB-10991 Allow to edit on Azure and GCP. 2 lists added, one for existed ssh key and one for raw ssh key. Only one supported by provider, cannot add the same provider to each lists.

+++ environment/src/main/java/com/sequenceiq/environment/environment/service/EnvironmentModificationService.java
@@ -210,20 +210,27 @@ public class EnvironmentModificationService {
                 throw new BadRequestException(validationResult.getFormattedErrors());
             }
             EnvironmentAuthentication originalAuthentication = environment.getAuthentication();
-            environment.setAuthentication(authenticationDtoConverter.dtoToAuthentication(authenticationDto));
-            boolean cleanupOldSshKey = true;
-            if (StringUtils.isNotEmpty(authenticationDto.getPublicKey())) {
-                cleanupOldSshKey = environmentResourceService.createAndUpdateSshKey(environment);
-            }
-            if (cleanupOldSshKey) {
-                String oldSshKeyId = originalAuthentication.getPublicKeyId();
-                LOGGER.info("The '{}' of ssh key is replaced with {}", oldSshKeyId, environment.getAuthentication().getPublicKeyId());
-                if (originalAuthentication.isManagedKey()) {
-                    environmentResourceService.deletePublicKey(environment, oldSshKeyId);
+            if (environmentResourceService.isRawSshKeyUpdateSupported(environment)) {
+                EnvironmentAuthentication updated = authenticationDtoConverter.dtoToSshUpdatedAuthentication(authenticationDto);
+                updated.setLoginUserName(originalAuthentication.getLoginUserName());
+                updated.setId(originalAuthentication.getId());
+                environment.setAuthentication(updated);
+            } else if (environmentResourceService.isExistingSshKeyUpdateSupported(environment)) {
+                environment.setAuthentication(authenticationDtoConverter.dtoToAuthentication(authenticationDto));
+                boolean cleanupOldSshKey = true;
+                if (StringUtils.isNotEmpty(authenticationDto.getPublicKey())) {
+                    cleanupOldSshKey = environmentResourceService.createAndUpdateSshKey(environment);
                 }
-            } else {
-                LOGGER.info("Authentication modification was unsuccessful. The authentication was reverted to the previous version.");
-                environment.setAuthentication(originalAuthentication);
+                if (cleanupOldSshKey) {
+                    String oldSshKeyId = originalAuthentication.getPublicKeyId();
+                    LOGGER.info("The '{}' of ssh key is replaced with {}", oldSshKeyId, environment.getAuthentication().getPublicKeyId());
+                    if (originalAuthentication.isManagedKey()) {
+                        environmentResourceService.deletePublicKey(environment, oldSshKeyId);
+                    }
+                } else {
+                    LOGGER.info("Authentication modification was unsuccessful. The authentication was reverted to the previous version.");
+                    environment.setAuthentication(originalAuthentication);
+                }
             }
         }
     }
